refactor: replace image generation API with new analyzeImage server action and update related components
