+++ .env.example
@@ -4,6 +4,5 @@ NEXT_PUBLIC_APP_URL=""
 
 API_KEY=""
 NEXT_PUBLIC_API_BASE_URL=""
-NEXT_PUBLIC_AI_MODEL=""
 
 GA_MEASUREMENT_ID=""
