+++ src/lib/types.ts
@@ -1,0 +1,7 @@
+export type GeminiAPIError = {
+  error: {
+    code: number;
+    message: string;
+    status: string;
+  };
+};
